+++ utils/src/main/kotlin/io/bkbn/stoik/utils/KotlinPoetUtils.kt
@@ -1,12 +1,15 @@
 package io.bkbn.stoik.utils
 
 import com.google.devtools.ksp.symbol.KSPropertyDeclaration
+import com.google.devtools.ksp.symbol.KSTypeReference
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.CodeBlock
 import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.ParameterSpec
 import com.squareup.kotlinpoet.PropertySpec
+import com.squareup.kotlinpoet.TypeName
 import com.squareup.kotlinpoet.ksp.KotlinPoetKspPreview
+import com.squareup.kotlinpoet.ksp.toClassName
 import com.squareup.kotlinpoet.ksp.toTypeName
 import io.bkbn.stoik.core.Domain
 
@@ -26,6 +29,17 @@ object KotlinPoetUtils {
     endControlFlow()
   }
 
+  fun CodeBlock.Builder.addObjectInstantiation(
+    type: TypeName,
+    init: CodeBlock.Builder.() -> Unit
+  ) {
+    add("%T(\n", type)
+    indent()
+    add(CodeBlock.Builder().apply(init).build())
+    unindent()
+    add(")\n")
+  }
+
   fun FunSpec.Builder.addCodeBlock(
     init: CodeBlock.Builder.() -> Unit
   ) {
@@ -38,12 +52,26 @@ object KotlinPoetUtils {
   fun Domain.toEntityClass(): ClassName = ClassName(BASE_ENTITY_PACKAGE_NAME, name.plus("Entity"))
   fun Domain.toDaoClass(): ClassName = ClassName(BASE_ENTITY_PACKAGE_NAME, name.plus("Dao"))
 
-  fun KSPropertyDeclaration.toParameter() =
-    ParameterSpec.builder(simpleName.getShortName(), type.toTypeName()).build()
+  fun String.toResponseClass(): ClassName = ClassName(BASE_MODEL_PACKAGE_NAME, this.plus("Response"))
+  fun String.toEntityClass(): ClassName = ClassName(BASE_ENTITY_PACKAGE_NAME, this.plus("Entity"))
+
+  fun ClassName.toEntityClass(): ClassName = ClassName(BASE_ENTITY_PACKAGE_NAME, simpleName.plus("Entity"))
+
+  fun KSPropertyDeclaration.toParameter() = ParameterSpec.builder(simpleName.getShortName(), type.toTypeName()).build()
 
   fun KSPropertyDeclaration.toProperty(isMutable: Boolean = false) =
     PropertySpec.builder(simpleName.getShortName(), type.toTypeName()).apply {
       if (isMutable) mutable()
       initializer(simpleName.getShortName())
     }.build()
+
+  fun KSTypeReference.isSupportedScalar(): Boolean = when (this.resolve().toClassName().simpleName) {
+    "String" -> true
+    "Int" -> true
+    "Long" -> true
+    "Double" -> true
+    "Float" -> true
+    "Boolean" -> true
+    else -> false
+  }
 }